SigilHive/pyproject.toml
[project]
name = "sigilhive"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncssh>=2.21.1",
    "langchain>=0.3.27",
    "langchain-google-genai>=2.1.12",
    "numpy>=2.3.3",
    "python-dotenv>=1.1.1",
]
